Refactor Elixir build and run scripts, adding escript support
The commit message summarizes the changes made in the commit. In this case, the commit message could be:

"Refactor Elixir build and run scripts, adding escript support"

This message conveys that the Elixir build and run scripts have been modified to use escript, allowing the code to be compiled into a standalone
executable.